add user_actions
